add shared hook
