+++ apps/src/codebridge/Settings/SettingsDropdown.tsx
@@ -5,11 +5,12 @@ import SimpleDropdown, {
 } from '@code-dot-org/component-library/dropdown/simpleDropdown';
 import {Heading6} from '@code-dot-org/component-library/typography';
 import FocusTrap from 'focus-trap-react';
-import React, {useLayoutEffect, useState} from 'react';
+import React, {useState} from 'react';
 import {createPortal} from 'react-dom';
 
 import codebridgeI18n from '@cdo/apps/codebridge/locale';
 import {DEFAULT_FONT_SIZE_KEY, FontSize} from '@cdo/apps/lab2/constants';
+import useDropdownPosition from '@cdo/apps/lab2/hooks/useDropdownPosition';
 import {setEditorFontSize} from '@cdo/apps/lab2/redux/lab2ViewRedux';
 import {SignInState} from '@cdo/apps/templates/currentUserRedux';
 import useOutsideClick from '@cdo/apps/util/hooks/useOutsideClick';
@@ -21,10 +22,8 @@ import {useCodebridgeContext} from '../codebridgeContext';
 
 import moduleStyles from './settings-dropdown.module.scss';
 
-const TOP_PADDING = 5;
 // fontSizeOptions contains a list of value/text from the FontSize enum,
 // e.g., [{value: 'Tiny', text: 'Tiny'}, {value: 'Small', text: 'Small'}, ...]
-
 const fontSizeOptions: SimpleDropdownProps['items'] = Object.keys(FontSize)
   .filter(key => isNaN(Number(key))) // Filters out the reverse enum keys.
   .map(key => ({
@@ -41,7 +40,6 @@ const SettingsDropdown: React.FunctionComponent<SettingsDropdownProps> = ({
   buttonRef,
 }) => {
   const menuRef = useOutsideClick<HTMLDivElement>(closeDropdown);
-  const [dropdownStyles, setDropdownStyles] = useState<React.CSSProperties>({});
   const editorFontSizeKey = useAppSelector(
     state => state.lab2View.editorFontSizeKey
   );
@@ -53,24 +51,7 @@ const SettingsDropdown: React.FunctionComponent<SettingsDropdownProps> = ({
 
   const getSelectedKey = (value: string) => value as keyof typeof FontSize;
 
-  useLayoutEffect(() => {
-    const updateDropdownPosition = () => {
-      if (buttonRef.current && menuRef.current) {
-        const dropdownRect = menuRef.current.getBoundingClientRect();
-        const parentRect = buttonRef.current.getBoundingClientRect();
-        const top =
-          parentRect.top + parentRect.height + TOP_PADDING + window.scrollY;
-        const left = parentRect.right - dropdownRect.width + window.scrollX;
-        setDropdownStyles({top, left});
-      }
-    };
-
-    updateDropdownPosition();
-    window.addEventListener('resize', updateDropdownPosition);
-    return () => {
-      window.removeEventListener('resize', updateDropdownPosition);
-    };
-  }, [buttonRef, menuRef]);
+  const dropdownStyles = useDropdownPosition(buttonRef, menuRef);
 
   const onTextEditorDropdownChange = async (value: string) => {
     setSelectedValue(getSelectedKey(value));
