+++ apps/src/lab2/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useFileUploader';
 export * from './useKeyboardTrap';
 export * from './useLifecycleNotifier';
 export * from './usePartialApply';
+export * from './useDropdownPosition';
